When a service is removed, also remove the charm archive if local
