+++ state/cleanup_test.go
@@ -12,8 +12,10 @@ import (
 	gc "gopkg.in/check.v1"
 	"gopkg.in/juju/charm.v6-unstable"
 
+	"bytes"
 	"github.com/juju/juju/instance"
 	"github.com/juju/juju/state"
+	"github.com/juju/juju/state/storage"
 	"github.com/juju/juju/storage/provider"
 	"github.com/juju/juju/storage/provider/registry"
 	"github.com/juju/juju/testing/factory"
@@ -70,6 +72,30 @@ func (s *CleanupSuite) TestCleanupDyingServiceUnits(c *gc.C) {
 	// Run a final cleanup to clear the cleanup scheduled for the unit that
 	// became dying.
 	s.assertCleanupCount(c, 1)
+}
+
+func (s *CleanupSuite) TestCleanupDyingServiceCharm(c *gc.C) {
+	// Create a service and  a charm.
+	ch := s.AddTestingCharm(c, "mysql")
+	mysql := s.AddTestingService(c, "mysql", ch)
+
+	// Create a dummy archive blob.
+	stor := storage.NewStorage(s.State.ModelUUID(), s.State.MongoSession())
+	storagePath := "dummy-path"
+	err := stor.Put(storagePath, bytes.NewReader([]byte("data")), 4)
+	c.Assert(err, jc.ErrorIsNil)
+
+	// Destroy the service and check that a cleanup has been scheduled.
+	err = mysql.Destroy()
+	c.Assert(err, jc.ErrorIsNil)
+	s.assertNeedsCleanup(c)
+
+	// Run the cleanup, and check that the charm is removed.
+	s.assertCleanupRuns(c)
+	_, _, err = stor.Get(storagePath)
+	c.Assert(err, jc.Satisfies, errors.IsNotFound)
+	_, err = s.State.Charm(ch.URL())
+	c.Assert(err, jc.Satisfies, errors.IsNotFound)
 }
 
 func (s *CleanupSuite) TestCleanupControllerModels(c *gc.C) {
